Finalise applicant attributes
